requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.12.0=pyhd8ed1ab_0
blas=1.0=mkl
ca-certificates=2021.4.13=h06a4308_1
certifi=2020.12.5=py38h06a4308_0
cudatoolkit=11.0.221=h6bb024c_0
cycler=0.10.0=py38_0
dbus=1.13.18=hb2f20db_0
expat=2.2.10=he6710b0_2
fontconfig=2.13.0=h9420a91_0
freetype=2.10.4=h5ab3b9f_0
glib=2.66.1=h92f7085_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
icu=58.2=he6710b0_3
intel-openmp=2020.2=254
jax=0.2.12=pyhd8ed1ab_0
jaxlib=0.1.64=py38h578d9bd_0
joblib=1.0.0=pyhd3eb1b0_0
jpeg=9b=h024ee3a_2
kiwisolver=1.3.0=py38h2531618_0
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuuid=1.0.3=h1bed415_2
libuv=1.40.0=h7b6447c_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=hb55368b_3
lz4-c=1.9.2=heb0550a_3
matplotlib=3.3.2=h06a4308_0
matplotlib-base=3.3.2=py38h817c723_0
mkl=2020.2=256
mkl-service=2.3.0=py38he904b0f_0
mkl_fft=1.2.0=py38h23d657b_0
mkl_random=1.1.1=py38h0573a6f_0
ncurses=6.2=he6710b0_1
ninja=1.10.2=py38hff7bd54_0
numpy=1.19.2=py38h54aff64_0
numpy-base=1.19.2=py38hfa32c7d_0
olefile=0.46=py_0
openssl=1.1.1k=h27cfd23_0
opt_einsum=3.3.0=py_0
pandas=1.2.4=py38h2531618_0
pcre=8.44=he6710b0_0
pillow=8.1.0=py38he98fc37_0
pip=20.3.3=py38h06a4308_0
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py38h05f1152_4
python=3.8.5=h7579374_1
python-dateutil=2.8.1=py_0
python-flatbuffers=1.12=pyhd8ed1ab_1
python_abi=3.8=1_cp38
pytorch=1.7.1=py3.8_cuda11.0.221_cudnn8.0.5_0
pytz=2021.1=pyhd3eb1b0_0
pyyaml=5.3.1=py38h7b6447c_1
qt=5.9.7=h5867ecd_1
readline=8.0=h7b6447c_0
scandir=1.10.0=pypi_0
scikit-learn=0.23.2=py38h0573a6f_0
scipy=1.5.2=py38h0b6359f_0
setuptools=51.3.3=py38h06a4308_4
sip=4.19.13=py38he6710b0_0
six=1.15.0=py38h06a4308_0
sqlite=3.33.0=h62c20be_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=hbc83047_0
torchaudio=0.7.2=py38
torchvision=0.8.2=py38_cu110
tornado=6.1=py38h27cfd23_0
typing_extensions=3.7.4.3=py_0
wheel=0.36.2=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
